fix occurrences of the WorkspaceCompiler rename
